requirements.txt
torch==2.2.2
torchvision==0.17.2
albumentations==1.4.4
tqdm==4.66.2
matplotlib==3.8.4
scikit-learn==1.4.2
hydra-core==1.3.2
numpy==1.26.4
opencv-python==4.9.0.80
terminaltables==3.1.10
